+++ src/identity_mapper_service/__init__.py
@@ -1,0 +1,10 @@
+"""IdentityMapper Host Service."""
+
+from identity_mapper_service.registry import ProviderRegistry, UnknownProviderError
+from identity_mapper_service.service import IdentityMapperHostService
+
+__all__ = [
+    "IdentityMapperHostService",
+    "ProviderRegistry",
+    "UnknownProviderError",
+]